Add download script for PyQUDA.

+++ cuda/build_offline.sh
@@ -7,7 +7,7 @@ LLVM_VERSION=16
 JOBS=32
 QUDA_JOBS=32
 
-ROOT=$HOME/scidac
+ROOT=$(cd $(dirname ${BASH_SOURCE[0]}) && pwd)
 SRC=${ROOT}
 BIN=${ROOT}/build
 DST=${ROOT}/install
@@ -38,10 +38,14 @@ function prepare_quda() {
     if [ $1 -gt 0 ]; then
         mkdir -p ${BIN}/$2
         pushd ${BIN}/$2
-        mkdir -p cmake
-        cp ${ROOT}/CPM_0.38.5.cmake cmake
-        mkdir -p _deps/eigen-subbuild/eigen-populate-prefix/src
-        cp ${ROOT}/eigen-3.4.0.tar.bz2 _deps/eigen-subbuild/eigen-populate-prefix/src
+        if !([ -f cmake/CPM_0.38.5.cmake ] && [ "$(sha256sum cmake/CPM_0.38.5.cmake | awk '{print $1}')" == "192aa0ccdc57dfe75bd9e4b176bf7fb5692fd2b3e3f7b09c74856fc39572b31c" ]); then
+            mkdir -p cmake
+            cp ${ROOT}/CPM_0.38.5.cmake cmake
+        fi
+        if !([ -f _deps/eigen-subbuild/eigen-populate-prefix/src/eigen-3.4.0.tar.bz2 ] && [ "$(sha256sum _deps/eigen-subbuild/eigen-populate-prefix/src/eigen-3.4.0.tar.bz2 | awk '{print $1}')" == "b4c198460eba6f28d34894e3a5710998818515104d6e74e5cc331ce31e46e626" ]); then
+            mkdir -p _deps/eigen-subbuild/eigen-populate-prefix/src
+            cp ${ROOT}/eigen-3.4.0.tar.bz2 _deps/eigen-subbuild/eigen-populate-prefix/src
+        fi
         popd
     fi
 }
@@ -120,6 +124,7 @@ build ${BUILD_PYQUDA} ${QUDA_JOBS} pyquda \
     cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DCMAKE_BUILD_TYPE=RELEASE -DQUDA_BUILD_SHAREDLIB=ON \
     -DQUDA_GPU_ARCH=${GPU_TARGET} -DQUDA_HETEROGENEOUS_ATOMIC=${HETEROGENEOUS_ATOMIC} \
     -DQUDA_MPI=ON \
+    -DQUDA_CONTRACT=ON -DQUDA_COVDEV=ON \
     -DQUDA_CLOVER_DYNAMIC=OFF -DQUDA_CLOVER_RECONSTRUCT=OFF \
     -DQUDA_DIRAC_CLOVER_HASENBUSCH=OFF -DQUDA_DIRAC_DOMAIN_WALL=OFF \
     -DQUDA_DIRAC_TWISTED_CLOVER=OFF -DQUDA_DIRAC_TWISTED_MASS=OFF -DQUDA_DIRAC_NDEG_TWISTED_CLOVER=OFF -DQUDA_DIRAC_NDEG_TWISTED_MASS=OFF \
